maken van #38 component
voor #37 heb ik verschillende componenten ingedeeld in een library. Bij deze ook de Typegraph
