+++ src/lib/index.js
@@ -1,14 +1,15 @@
 // place files you want to import through the `$lib` alias in this folder.
 export { default as Icon } from './1_Atoms/Icon.svelte'
+
+export { default as BarChart } from './2_Molecules/BarChart.svelte'
 export { default as DonutChart } from './2_Molecules/DonutChart.svelte'
 export { default as MonthlyNavigation } from './2_Molecules/MonthlyNavigation.svelte'
 
-
-export { default as Title } from './4_Templates/Title.svelte'
-export { default as AccessibillityGraph } from './4_Templates/AccessibillityGraph.svelte'
-export { default as AutomaticScan } from './4_Templates/AutomaticScan.svelte'
-export { default as ErrorResult } from './4_Templates/ErrorResult.svelte'
-export { default as TypeGraph } from './4_Templates/TypeGraph.svelte'
-export { default as Header } from './4_Templates/Header.svelte'
-export { default as Footer } from './4_Templates/Footer.svelte'
+export { default as Title } from './3_Organism/Title.svelte'
+export { default as AccessibillityGraph } from './3_Organism/AccessibillityGraph.svelte'
+export { default as AutomaticScan } from './3_Organism/AutomaticScan.svelte'
+export { default as ErrorResult } from './3_Organism/ErrorResult.svelte'
+export { default as TypeGraph } from './3_Organism/TypeGraph.svelte'
+export { default as Header } from './3_Organism/Header.svelte'
+export { default as Footer } from './3_Organism/Footer.svelte'
